backend/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
python-dotenv
pinecone
google-generativeai